+++ apps/server/src/routes/pool.ts
@@ -293,12 +293,12 @@ export async function poolRoutes(app: FastifyInstance) {
         const supplyResult = await c.query(
           `UPDATE app_counters
              SET value = value + (CASE
-               WHEN name = 'minted_supply' THEN $1::bigint
-               ELSE 1::bigint
+               WHEN name = 'minted_supply' THEN $1::numeric
+               ELSE 1
              END)
            WHERE asset_id=$3::uuid
              AND name IN ('minted_supply','block_height')
-             AND ($4::boolean OR (SELECT value FROM app_counters WHERE asset_id=$3::uuid AND name='minted_supply') + $1::bigint <= $2::bigint)`,
+             AND ($4::boolean OR (SELECT value FROM app_counters WHERE asset_id=$3::uuid AND name='minted_supply') + $1::numeric <= $2::numeric)`,
           [reward.toString(), capBaseUnits.toString(), assetId, String(asset.supplyMode === 'unlimited')],
         );
         if (supplyResult.rowCount !== 2) {
@@ -389,7 +389,7 @@ export async function poolRoutes(app: FastifyInstance) {
         // The full gross reward enters circulation here, regardless of
         // how the per-recipient MINT events split it below.
         await c.query(
-          `UPDATE ledger_stats SET value = value + $1::bigint, updated_at = now()
+          `UPDATE ledger_stats SET value = value + $1::numeric, updated_at = now()
            WHERE asset_id=$2::uuid AND name='circulating_supply'`,
           [grossReward.toString(), asset.id],
         );
@@ -473,7 +473,7 @@ export async function poolRoutes(app: FastifyInstance) {
 
           await c.query(
             `INSERT INTO pool_payouts(asset_id, round_id, pubkey, share_count, payout_base_units, is_finder, event_id)
-             VALUES($1::uuid, $2::bigint, $3, $4::bigint, $5::bigint, $6, $7)`,
+             VALUES($1::uuid, $2::bigint, $3, $4::bigint, $5::numeric, $6, $7)`,
             [assetId, roundId, recipient, shareCount.toString(), amount.toString(), isFinder, eventId],
           );
           return eventId;
@@ -568,10 +568,10 @@ export async function poolRoutes(app: FastifyInstance) {
              SET ended_at = now(),
                  ended_by_pubkey = $2,
                  ended_by_event_id = $3,
-                 reward_base_units = $4::bigint,
-                 treasury_cut_base_units = $5::bigint,
-                 finder_payout_base_units = $6::bigint,
-                 pro_rata_pool_base_units = $7::bigint,
+                 reward_base_units = $4::numeric,
+                 treasury_cut_base_units = $5::numeric,
+                 finder_payout_base_units = $6::numeric,
+                 pro_rata_pool_base_units = $7::numeric,
                  participant_count = $8
            WHERE asset_id=$9::uuid AND id = $1::bigint`,
           [
